feat(appliance): backport all recent appliance changes (#64182)
Draft in case plan in
https://linear.app/sourcegraph/issue/REL-309/release-process-for-appliance
not agreed. Please see that first.


```
git log --format=%H d47b4cc48b6ea27cf6b5a274b79a6a4c8f38cf8c..origin/main -- cmd/appliance internal/appliance docker-images/appliance-frontend | tac | xargs git cherry-pick
```

d47b4cc48b6ea27cf6b5a274b79a6a4c8f38cf8c being the commit we branched
off main from to create the 5.5.x branch
(https://buildkite.com/sourcegraph/sourcegraph/builds/281882).

Commits (generated by `git log --format='-
https://github.com/sourcegraph/sourcegraph/commit/%H'
d47b4cc48b6ea27cf6b5a274b79a6a4c8f38cf8c..origin/main -- cmd/appliance
internal/appliance docker-images/appliance-frontend | tac`):

-
https://github.com/sourcegraph/sourcegraph/commit/a20b0650b453d11d7dc428b9f919b79465db088c
-
https://github.com/sourcegraph/sourcegraph/commit/b71c986c77829cfe3bdc6f2d31e8590fdb969b17
-
https://github.com/sourcegraph/sourcegraph/commit/91864283bc0811928188f415e6cda7d3fe940e34
-
https://github.com/sourcegraph/sourcegraph/commit/c88b57020fa49fe996e37ca9409a541d48d7a63e
-
https://github.com/sourcegraph/sourcegraph/commit/0491839942c3087687d0f48ade1c544f1fbcb881
-
https://github.com/sourcegraph/sourcegraph/commit/619fc5707415bd50631153bcb1b01c670d44fd82
-
https://github.com/sourcegraph/sourcegraph/commit/e81c39a834810a60c8827456963511213eb5f602
-
https://github.com/sourcegraph/sourcegraph/commit/a61f353e0e93f848fa476ffda6d46f71060d465d
-
https://github.com/sourcegraph/sourcegraph/commit/0abef7b43d5dadac92a3ff318f3caf6d7b572f6a
-
https://github.com/sourcegraph/sourcegraph/commit/0e391a964ae8d21904bd875862759db521feb5ba
-
https://github.com/sourcegraph/sourcegraph/commit/daae9adfb642f2c73e7b4b2dcb7acca618ae86d3
-
https://github.com/sourcegraph/sourcegraph/commit/6e31f0f4cc89ad7d3892e67bad55ab9a0b9658a7
-
https://github.com/sourcegraph/sourcegraph/commit/49a600220d5eb2875ab54c87329d0a90f15c38d6
-
https://github.com/sourcegraph/sourcegraph/commit/37cf4a7b7e9dfb4f09635d2e04ce30daf5d8b7a0
-
https://github.com/sourcegraph/sourcegraph/commit/29fc613c376fe38055fa907c5bbd293ea6e4408f
-
https://github.com/sourcegraph/sourcegraph/commit/255e6387cc7854c8aed6dfa1445fcade380ec690
-
https://github.com/sourcegraph/sourcegraph/commit/49b32fcf3a1ec8362e3be286de182aba76f93917
-
https://github.com/sourcegraph/sourcegraph/commit/9f4c160f91838ed3167a38ba62b7ca90ad90a771
-
https://github.com/sourcegraph/sourcegraph/commit/3814fd7390455354be1129cdb1b240c201ef5964
-
https://github.com/sourcegraph/sourcegraph/commit/c68e92bc28b3ce0878aebe8e10471cca89010b77
-
https://github.com/sourcegraph/sourcegraph/commit/7e82c27ab5de1c76f1e877ec9dee665800807cbb
-
https://github.com/sourcegraph/sourcegraph/commit/98c6b9703f889bb8b706855204276d3528b03356
-
https://github.com/sourcegraph/sourcegraph/commit/a01ebad8417d4c14b969f61a3a6ad8e836728047
-
https://github.com/sourcegraph/sourcegraph/commit/8c2d8da234cc46b529b8df92a38cdf7f7d83a7a7
-
https://github.com/sourcegraph/sourcegraph/commit/ebec72d7ed7046406d22d2bf8276612945db358d
-
https://github.com/sourcegraph/sourcegraph/commit/d945f192852a7dc3cd2482602fe37f614f82f8a1
-
https://github.com/sourcegraph/sourcegraph/commit/84e28998e9d3976b79d858439953d101ca44ab7e


## Test plan

Tests pass.

## Changelog

- Backport all recent appliance changes. The appliance is still
pre-release.

---------

+++ internal/appliance/reconciler/kubernetes.go
@@ -11,6 +11,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/sourcegraph/sourcegraph/internal/appliance/config"
@@ -28,7 +29,7 @@ func reconcileObject[T client.Object](
 	sg *config.Sourcegraph, owner client.Object,
 ) error {
 	if cfg.IsDisabled() {
-		return r.ensureObjectDeleted(ctx, obj)
+		return ensureObjectDeleted(ctx, r, owner, obj)
 	}
 
 	updateIfChanged := struct {
@@ -62,7 +63,11 @@ func createOrUpdateObject[R client.Object](
 	ctx context.Context, r *Reconciler, updateIfChanged any,
 	owner client.Object, obj, objKind R,
 ) error {
-	logger := log.FromContext(ctx).WithValues("kind", obj.GetObjectKind().GroupVersionKind(), "namespace", obj.GetNamespace(), "name", obj.GetName())
+	gvk, err := apiutil.GVKForObject(obj, r.Scheme)
+	if err != nil {
+		return errors.Wrap(err, "getting GVK for object")
+	}
+	logger := log.FromContext(ctx).WithValues("kind", gvk.String(), "namespace", obj.GetNamespace(), "name", obj.GetName())
 	namespacedName := types.NamespacedName{Namespace: obj.GetNamespace(), Name: obj.GetName()}
 
 	cfgHash, err := configHash(updateIfChanged)
@@ -81,7 +86,7 @@ func createOrUpdateObject[R client.Object](
 	// error: "cluster-scoped resource must not have a namespace-scoped owner".
 	// non-namespaced resources will therefore not be garbage-collected when the
 	// ConfigMap is deleted.
-	if !isNamespaced(obj) {
+	if isNamespaced(obj) {
 		if err := ctrl.SetControllerReference(owner, obj, r.Scheme); err != nil {
 			return errors.Newf("setting controller reference: %w", err)
 		}
@@ -102,6 +107,11 @@ func createOrUpdateObject[R client.Object](
 		return err
 	}
 
+	if !isControlledBy(owner, existingRes) && isNamespaced(obj) && !config.ShouldAdopt(obj) {
+		logger.Info("refusing to update non-owned resource")
+		return nil
+	}
+
 	if cfgHash != existingRes.GetAnnotations()[config.AnnotationKeyConfigHash] {
 		logger.Info("Found existing object with spec that does not match desired state. Clobbering it.")
 		if err := r.Client.Update(ctx, obj); err != nil {
@@ -117,18 +127,40 @@ func createOrUpdateObject[R client.Object](
 
 func isNamespaced(obj client.Object) bool {
 	if _, ok := obj.(*rbacv1.ClusterRole); ok {
-		return true
+		return false
 	}
 	if _, ok := obj.(*rbacv1.ClusterRoleBinding); ok {
-		return true
+		return false
 	}
-	return false
+	return true
 }
 
-func (r *Reconciler) ensureObjectDeleted(ctx context.Context, obj client.Object) error {
-	logger := log.FromContext(ctx).WithValues("kind", obj.GetObjectKind().GroupVersionKind(), "namespace", obj.GetNamespace(), "name", obj.GetName())
+func ensureObjectDeleted[T client.Object](ctx context.Context, r *Reconciler, owner client.Object, obj T) error {
+	// We need to try to get the object first, in order to check its owner
+	// references later.
+	objKey := types.NamespacedName{Name: obj.GetName(), Namespace: obj.GetNamespace()}
+	if err := r.Client.Get(ctx, objKey, obj); err != nil {
+		if kerrors.IsNotFound(err) {
+			// Object doesn't exist, we don't need to delete it
+			return nil
+		}
+	}
+	gvk, err := apiutil.GVKForObject(obj, r.Scheme)
+	if err != nil {
+		return errors.Wrap(err, "getting GVK for object")
+	}
+
+	logger := log.FromContext(ctx).WithValues("kind", gvk.String(), "namespace", obj.GetNamespace(), "name", obj.GetName())
+
+	if !isControlledBy(owner, obj) && isNamespaced(obj) {
+		logger.Info("refusing to delete non-owned resource")
+		return nil
+	}
+
+	logger.Info("deleting resource")
 	if err := r.Client.Delete(ctx, obj); err != nil {
 		if kerrors.IsNotFound(err) {
+			// If by chance it got deleted concurrently, no harm done.
 			return nil
 		}
 
@@ -136,6 +168,15 @@ func (r *Reconciler) ensureObjectDeleted(ctx context.Context, obj client.Object)
 		return err
 	}
 	return nil
+}
+
+func isControlledBy(owner, obj client.Object) bool {
+	for _, ownerRef := range obj.GetOwnerReferences() {
+		if owner.GetUID() == ownerRef.UID {
+			return true
+		}
+	}
+	return false
 }
 
 func configHash(configElement any) (string, error) {
